requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.2
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==3.0.1
cryptography==41.0.4
Deprecated==1.2.14
discord==2.1.0
discord.py==2.1.1
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
pycparser==2.21
PyGithub==2.1.1
PyJWT==2.8.0
PyNaCl==1.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.28.2
six==1.16.0
soupsieve==2.4
typing_extensions==4.8.0
urllib3==1.26.14
wrapt==1.15.0
yarl==1.8.2
